+++ tests/test_simulations.py
@@ -19,7 +19,8 @@ class TestSimulation():
         receivers = (np.arange(12)*500, 0, -1000, 0, 0)
         frequencies = (1.0, 2.0)
 
-        survey = surveys.Survey('Test', sources, receivers, frequencies)
+        survey = surveys.Survey('Test', sources, receivers, frequencies,
+                                noise_floor=1e-15, relative_error=0.05)
 
         # Create a simple grid and model
         grid = meshes.TensorMesh(
@@ -36,7 +37,6 @@ class TestSimulation():
 
         # Do first one single and then all together.
         simulation.get_efield('Tx0', 2.0)
-        simulation.compute(reference=True)
 
     def test_derived(self):